move pointing device code into userspace
